transitionfunction now returns an int, which is the Numbered State, that the CA will look up in it's states array instead of "closest" checking

+++ CACPPN/CPPNNEAT/CA/Base/BaseNeatCA.cs
@@ -34,7 +34,7 @@ namespace CPPNNEATCA.CA.Base
 			return state;
 		}
 
-		public abstract float RunEvaluation(Func<List<float>, float> TransitionFunction);
+		public abstract float RunEvaluation(Func<List<float>, int> TransitionFunction);
 
 		public CAParameters GetParameters()
 		{
