+++ CACPPN/CPPNNEAT/CA/Experiments/Test1DimCA.cs
@@ -36,7 +36,7 @@ namespace CPPNNEATCA.CA.Experiments
 				cell.InitializeNeighbourhood();
 		}
 
-		public override float RunEvaluation(Func<List<float>, float> TransitionFunction)
+		public override float RunEvaluation(Func<List<float>, int> TransitionFunction)
 		{
 			Console.WriteLine();
 			Thread.Sleep(300);
